Backport PR from Beam: Move copy and remove from FileBasedSink to GcsUtil.
